pyproject.toml
[project]
name = "busca-eventos-rio"
version = "0.1.0"
description = "Sistema multi-agente para busca de eventos culturais no Rio de Janeiro"
requires-python = ">=3.11"
dependencies = [
    "agno>=2.0.0",
    "openai>=1.58.1",
    "duckduckgo-search>=6.3.11",
    "beautifulsoup4>=4.12.3",
    "playwright>=1.49.1",
    "httpx>=0.28.1",
    "pydantic>=2.10.5",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.0.1",
    "lxml>=5.3.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "ruff>=0.9.1",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "ruff>=0.9.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
